Relocate deprecated typography API code
