+++ libs/@guardian/source-foundations/src/deprecated-exports.ts
@@ -35,4 +35,47 @@ export {
 	focus,
 } from './_deprecated/colour/palette';
 
-export type { TypographySizes } from './_deprecated/typography/types';
+export {
+	titlepiece,
+	headline,
+	body,
+	textSans,
+	titlepieceSizes,
+	headlineSizes,
+	bodySizes,
+	textSansSizes,
+	remTitlepieceSizes,
+	remHeadlineSizes,
+	remBodySizes,
+	remTextSansSizes,
+	fonts,
+	fontWeights,
+	lineHeights,
+} from './_deprecated/typography';
+
+export {
+	body as bodyObjectStyles,
+	headline as headlineObjectStyles,
+	textSans as textSansObjectStyles,
+	titlepiece as titlepieceObjectStyles,
+} from './_deprecated/typography/obj';
+
+export type {
+	ScaleUnit,
+	Category,
+	LineHeight,
+	FontWeight,
+	FontStyle,
+	FontWeightDefinition,
+	Option,
+	TypographySizes,
+	TypographyStyles,
+	TitlepieceSizes,
+	HeadlineSizes,
+	BodySizes,
+	TextSansSizes,
+	Fs,
+	FontScaleFunction,
+	FontScaleFunctionStr,
+	FontScaleArgs,
+} from './_deprecated/typography/types';
